move restore.go into its own sub-package
